+++ types/interfaces/parachains/definitions.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'node:buffer';
+
 
 /* eslint-disable sort-keys */